feat: HUDView Money Type int로 변경

+++ CollabProject-5Team/Assets/Project/Scripts/UI/HUDView.cs
@@ -60,7 +60,7 @@ namespace GameDevTycoon.UI.Ingame
             _timeLabel.text = $"{week}주차 {dayName} {(isNight ? "밤" : "낮")}";
         }
 
-        public void SetMoneyLabel(long money)
+        public void SetMoneyLabel(int money)
         {
             _moneyLabel.text = $"{money:N0}G";
         }
